hotfix

+++ _print_number.c
@@ -33,7 +33,7 @@ void _print_int(va_list args, int *count)
  */
 void _print_sint(va_list args, int *count)
 {
-	long int n = va_arg(args, long int);
+	int n = va_arg(args, int);
 
 	if (n >= 0)
 	{
